Początki obsługi Dictionary dla listy paczek w kontenerze

+++ Package master/Form1.cs
@@ -15,7 +15,8 @@ namespace Package_master
     public partial class Form1 : Form
     {
         private Graphics gRysuj;
-        private List<Package> Paczki = new List<Package>();
+        private List<Package> Packages = new List<Package>();
+        private Dictionary<int,Package> Packages_in_container = new Dictionary<int,Package>();
         public Form1()
         {
             InitializeComponent();
@@ -26,21 +27,19 @@ namespace Package_master
             Weight_numericUpDown.Minimum = 40;
             Weight_numericUpDown.Maximum = 400;
             Weight_numericUpDown.Value = 80;
-
         }
-
+        
 
         private void button1_Click_1(object sender, EventArgs e)
-        {
-           
-            Pen pen = new Pen(Color.Black);
-            int i = listBox1.SelectedIndex;
+        {           
+            Pen pen = new Pen(Color.Tomato);
+            
+            int i = lPackage_list.SelectedIndex;
             if (i > -1)
             {
-                gRysuj.DrawRectangle(pen,Paczki[i].Size);
+                gRysuj.DrawRectangle(pen, Packages[i].Size);
             }
-           
-
+          
         }
 
         private void comboBox1_SelectedIndexChanged(object sender, EventArgs e)
@@ -53,7 +52,7 @@ namespace Package_master
             bool can_add = true;
             Package temp = new Package((int)Height_numeric_updown.Value, (int)Weight_numericUpDown.Value);
             
-            foreach (Package t in Paczki)
+            foreach (Package t in Packages)
             {
                 if (t.ToString() == temp.ToString())
                 {
@@ -64,16 +63,13 @@ namespace Package_master
 
             if (can_add)
             {
-                Paczki.Add(temp);
-                listBox1.Items.Add(temp.ToString());
+                Packages.Add(temp);
+                lPackage_list.Items.Add(temp.ToString());
             }   
             
         }
 
-        private void listBox1_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            
-        }
+     
 
         private void numericUpDown1_ValueChanged(object sender, EventArgs e)
         {
@@ -92,32 +88,38 @@ namespace Package_master
 
         private void BDelete_Package_Click(object sender, EventArgs e)
         {
-            int i = listBox1.SelectedIndex;
+            int i = lPackage_list.SelectedIndex;
             if (i > -1)
             {
-                Paczki.RemoveAt(i);
-                listBox1.Items.RemoveAt(i);
+                Packages.RemoveAt(i);
+                lPackage_list.Items.RemoveAt(i);
             }
         }
 
         private void bSave_Click(object sender, EventArgs e)
         {
-            StreamWriter sw;
-            if (File.Exists("Package_list.txt"))
+            if (Packages.Count > 0)
             {
-                File.Delete("Package_list.txt");
-            }
-           
-            sw = File.CreateText("Package_list.txt");
-                foreach (Package t in Paczki)
+                StreamWriter sw;
+                if (File.Exists("Package_list.txt"))
+                {
+                    File.Delete("Package_list.txt");
+                }
+
+                sw = File.CreateText("Package_list.txt");
+                foreach (Package t in Packages)
                 {
                     sw.WriteLine(t.Size.Width);
                     sw.WriteLine(t.Size.Height);
                 }
-            sw.Close();
-               
-            
+                sw.Close();
+            }
+            else
+            {
+                MessageBox.Show("Lista jest pusta");
+            }
         }
+        
 
         private void bLoad_Click(object sender, EventArgs e)
         {
@@ -125,28 +127,32 @@ namespace Package_master
             if (File.Exists("Package_list.txt"))
             {
                 String[] tablica;
-                //// String x = File.ReadAllText("Package_list.txt");
-                // listBox1.Items.Add(x);
                 tablica = File.ReadAllLines("Package_list.txt");
-                Paczki.Clear();
-                listBox1.Items.Clear();
-                try
+                if (tablica.Length > 1)
                 {
-                    for (int i = 0; i < tablica.Length; i += 2)
+                    Packages.Clear();
+                    lPackage_list.Items.Clear();
+                    try
                     {
-                        int w = Int32.Parse(tablica[i]);
-                        int h = Int32.Parse(tablica[i + 1]);
-                        Package temp = new Package(w, h);
-                        Paczki.Add(temp);
-                        listBox1.Items.Add(temp);
-                    }
+                        for (int i = 0; i < tablica.Length; i += 2)
+                        {
+                            int h = Int32.Parse(tablica[i]);
+                            int w = Int32.Parse(tablica[i + 1]);
+                            Package temp = new Package(w, h);
+                            Packages.Add(temp);
+                            lPackage_list.Items.Add(temp);
+                        }
 
+                    }
+                    catch (FormatException)
+                    {
+                        MessageBox.Show("Plik jest uszkodzony!");
+                        File.Delete("Package_list.txt");
+                    }
                 }
-                catch (FormatException)
+                else
                 {
-                    MessageBox.Show("Plik jest uszkodzony!");
-                    File.Delete("Package_list.txt");
-                    //sw.Close();
+                    MessageBox.Show("Plik jest pusty. Zapisz wypełnioną listę");
                 }
             }
             else
@@ -155,6 +161,34 @@ namespace Package_master
             }
 
             
+        }
+
+        private void lPackage_list_SelectedIndexChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void bContainerAdd_Click(object sender, EventArgs e)
+        {
+            int i = lPackage_list.SelectedIndex;
+
+            Package temp = Packages[i];
+            if (Packages_in_container.ContainsValue(temp))
+            {
+                MessageBox.Show("ss");
+            }
+            else
+            {
+                Packages_in_container.Add(1, temp);
+            }
+
+            //if (Packages_in_container.ContainsValue(temp))
+            //{
+            //    MessageBox.Show("To już je");
+
+            //}
+            //Packages_in_container.Add(1, temp);
+            //lContainer_packages.Items.Add(temp.ToString() + "1fgh");
         }
     }
 }
